changed to reflect changes on bakcend, as usual

+++ src/testclient/TimedPoint2D.java
@@ -23,7 +23,7 @@ public class TimedPoint2D{
 		return latitude;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
@@ -31,7 +31,7 @@ public class TimedPoint2D{
 		return longitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
